fixing positional argument error - change dtype of image

+++ colorization.py
@@ -75,7 +75,7 @@ def optimize_latent_codes(args):
         img = imageio.imread(os.path.join(args.imgs_dir, img_name))
         img = cv2.resize(img, dsize=tuple(args.input_img_size))
 
-        img = tf.image.convert_image_dtype(img, dtype=tf.int32, saturate=False)
+        img = tf.image.convert_image_dtype(img, dtype=tf.float32, saturate=False)
 
         corrupted_img = convert_to_greyscale(img)
